+++ src/s_nk_1.c
@@ -9,6 +9,31 @@
 // stirling numbers of first kind
 // [ n ]
 // [ k ]
+//
+// s1(n, k) is the number of permutations of sets of size n with k cycles
+//
+// (1) recursive formula:
+//  s1(n, k) = {
+//    s1(n - 1, k - 1) + (n - 1) * s1(n - 1, k)
+//    0 | k == 0
+//    0 | k > n
+//    1 | n == 0, n == 0
+//  }
+//
+// (2) iterative formula:
+//  using tail recursion, we transform the "path" of calculating the s1(n, k)
+//  into a loop (zik-zak loop) like this:
+//  1 - - - - - - - - - - -
+//  1 1 - - - - - - - - - -
+//  1 * 1 - - - - - - - - -
+//  1 * * 1 - - - - - - - -
+//  1 * * * 1 - - - - - - -
+//  1 - * * * 1 - - - - - -
+//  1 - - * * - 1 - - - - -
+//  1 - - - x - - 1 - - - -
+//  1 - - - - - - - 1 - - -
+//  Therefore, to compute x, we only need to store and compute
+//  the elements of the table denoted by '*'.
 
 void calc_stirling_nk1(st_nk_t *st) {
 	const size_t n_k = st->n - st->k + 1;
@@ -34,8 +59,8 @@ main(const argc, char *argv[]) {
 		return EXIT_FAILURE;
 
 	size_t
-		n = atoi(argv[1]),
-		k = atoi(argv[2]);
+		n = atol(argv[1]),
+		k = atol(argv[2]);
 
 	if(k > n)
 		return EXIT_FAILURE;
